requirements.txt
streamlit==1.26.0
opencv-python-headless==4.8.0.76
face_recognition==1.3.0
matplotlib==3.7.0
numpy==1.23.5
pillow==9.4.0

